Return ERROR_GEN_FAILURE instead of false or -1. (I could not found a more really suitable error code, sorry.)
[ 1737922 ] CeReg functions return ERROR_SUCCESS on failure
Submitted By:
Ian Abbott - ijabbott

CeReg functions return ERROR_SUCCESS on failure
Most of the CeReg... functions in src/rapi/registry.c or
src/rapi2/registry2.c return 'false' when the call to 'rapi_context_call'
or 'rapi2_context_call' fails. The caller interprets this as
'ERROR_SUCCESS'. The exception to the rule is '_CeRegQueryValueEx' and
'_CeRegQueryValueEx2' which return '-1' on failure.

Most of the functions don't bother checking the result of the
'rapi_buffer_read_uint32' calls that read the remote error code and return
value, but '_CeRegSetValueEx' and '_CeRegSetValueEx2' functions return
'false' in this case (again interpreted as 'ERROR_SUCCESS'), whereas
'_CeRegQueryValueEx' and 'CeRegQueryValueEx2' return '-1' in this case.



git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@3024 599ab11a-0111-0410-8a50-d7a5b0953570